forwarding: extract local-delivery session caching into local_delivery.rs
Pure code-motion split of forwarding/mod.rs (#5650). Move the
host-local delivery session-miss caching path
(should_cache_local_delivery_session_on_miss,
install_helper_local_session_on_miss, ingress_interface_local_resolution*)
and the LOCAL_DELIVERY_IFINDEX0 diagnostic counter into
forwarding/local_delivery.rs.

Functions are moved verbatim; no logic, signature, or hot-path change.
pub(super) becomes pub(in crate::afxdp) so the mod.rs glob re-export
keeps all call sites resolving.

Validation: cargo build clean (0 errors).
